Create SQL Joins Exercise
Exercise that helps explore the use of tables and how they relate to each other in SQL using relationships. There are many types of relationships that explain how data in one table relates to data in another table. This is practice to become more familar with SQL Join methods (Many-to-Many Joins, Many-to-Many Outer Joins, Many-to-Many Insert)